src/chatbot/backend/pyproject.toml
[project]
name = "rag-book-assistant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "openai>=2.14.0",
    "openai-agents>=0.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]
